+++ app/Phoityne/VSCode/TH/EvaluateArgumentsJSON.hs
@@ -26,7 +26,7 @@ data EvaluateArguments =
     
     etc. 
   -}
-  , contextEvaluateArguments    :: String
+  , contextEvaluateArguments    :: Maybe String
     } deriving (Show, Read, Eq)